requirements.txt
requests>=2.31.0
psutil>=5.9.0
bcrypt>=4.1.2
PyYAML>=6.0.2
